pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "context-lens"
version = "0.1.0"
description = "Diagnose context degradation in LLM agents — find where memory breaks and how to fix it"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Rishi Bethi", email = "rishibethi2001@gmail.com" }]
keywords = ["llm", "agents", "langchain", "langgraph", "testing", "context-window", "rag", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.52.0",
    "tiktoken>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.2.0", "langchain-core>=0.3.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/Rishi-Bethi-007/context-lens"
Repository = "https://github.com/Rishi-Bethi-007/context-lens"
Issues = "https://github.com/Rishi-Bethi-007/context-lens/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=context_lens --cov-report=term-missing"
